requirements.txt
openai>=1.40.0
datasets>=2.20.0
tqdm>=4.66.0
pandas>=2.2.0
python-dotenv>=1.0.1
pyyaml>=6.0.1
tenacity>=8.2.3